+++ verifier-service/src/main/java/ch/admin/bj/swiyu/verifier/service/trustregistry/TrustStatementValidator.java
@@ -79,8 +79,9 @@ public class TrustStatementValidator {
         log.debug("Verifying trust statement signature for DID: {}", didString);
 
         // See: https://jira.bit.admin.ch/browse/EIDOMNI-959 - Architectural Risk: potential DoS via expensive DID resolution in signature verification
-        var didDoc = didResolverAdapter.resolveDid(didString, urlRewriteProperties.getUrlMappings());
-        trustStatementDidJwtValidator.validateJwt(jwtString, didDoc);
+        // For the time being deactivated
+        // var didDoc = didResolverAdapter.resolveDid(didString, urlRewriteProperties.getUrlMappings());
+        // trustStatementDidJwtValidator.validateJwt(jwtString, didDoc);
         log.debug("Trust statement signature verification succeeded for DID: {}", didString);
     }
 }